api/requirements.txt
asgiref==3.3.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.1.13
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-extra-fields==3.1.1
freezegun==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
install==1.3.4
mongoengine==0.23.0
postgis==1.0.4
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.4.0
